fix: next:goのjsonデータ名に合わせた(一部)、ディレクトリ構成 #37
