+++ t0016Next/myapp/src/pages/crud/edit.tsx
@@ -3,16 +3,15 @@ import { useRouter } from "next/router";
 import Link from 'next/link';
 import Router, { useEffect, useState } from 'react';
 // import style from '../Youtube.module.css';
-import type { CrudDate, ReceivedVtuber, ReceivedMovie, ReceivedKaraoke } from '../types/vtuber_content'; //type{}で型情報のみインポート
+import type { CrudDate, ReceivedVtuber, ReceivedMovie, ReceivedKaraoke } from '../../types/vtuber_content'; //type{}で型情報のみインポート
 // import DeleteButton from '../components/DeleteButton';
 import https from 'https';
 import axios from 'axios';
 import '@szhsin/react-menu/dist/index.css';
-import { DropDownVt, DropDownMo, DropDownKa } from '../components/Dropdown';
-import { YoutubePlayer } from '../components/YoutubePlayer'
-import { ConversionTime, ExtractVideoId } from '../components/Conversion'
-import { domain } from '../../env'
-
+import { DropDownVt, DropDownMo, DropDownKa } from '../../components/Dropdown';
+import { YouTubePlayer } from '../../components/YoutubePlayer'
+import { ConvertStringToTime, ExtractVideoId } from '../../components/Conversion'
+import { domain } from '../../../env'
 
 type TopPagePosts = {
     //   alljoindata: AllJoinData[];
@@ -40,10 +39,10 @@ export const AllDatePage = ({ posts, checkSignin }: AllDatePage) => {
             // setKaraokeStart(4)
         }
         if (selectedVtuber && selectedMovie) {
-            const foundMovie = posts.movies.find(movies => movies.Movie.MovieUrl === selectedMovie);
+            const foundMovie = posts.movies.find(movies => movies.MovieUrl === selectedMovie);
             // console.log("foundMovieUrl",foundMovie?.MovieUrl);
             if (foundMovie) {
-                const foundYoutubeId = ExtractVideoId(foundMovie.Movie.MovieUrl);
+                const foundYoutubeId = ExtractVideoId(foundMovie.MovieUrl);
                 setfoundMovie(foundYoutubeId);
                 setKaraokeStart(1)
                 //   console.log("foundYoutubeId", foundYoutubeId)
@@ -54,15 +53,17 @@ export const AllDatePage = ({ posts, checkSignin }: AllDatePage) => {
     useEffect(() => {
         if (selectedVtuber && selectedMovie && selectedKaraoke) {
             const foundMovies = posts.karaokes.filter(karaoke => karaoke.MovieUrl === selectedMovie);
-            const foundKaraoke = foundMovies.find(foundMovie => foundMovie.Movie.KaraokeId === selectedKaraoke)
+            const foundKaraoke = foundMovies.find(foundMovie => foundMovie.KaraokeId === selectedKaraoke)
+            // console.log("posts.karaokes", posts.karaokes) //karaoke_listsテーブルの全データをオブジェクトの配列で
+            // console.log("foundMovies", foundMovies)     //どういつURLのオブジェクトの配列
+            // console.log("foundKaraoke", foundKaraoke);  //karaokeの配列
 
             if (foundKaraoke) {
-                const foundSingStart = ConversionTime(foundKaraoke.SingStart);
+                const foundSingStart = ConvertStringToTime(foundKaraoke.SingStart);
                 setKaraokeStart(foundSingStart);
                 //   console.log("foundSingStart", foundSingStart)
             }
         }
-        console.log("selectedKaraoke=", selectedKaraoke)
     }, [selectedVtuber, selectedMovie, selectedKaraoke, posts.karaokes]);
 
     // 親選択クリア時に子もクリアするuseEffect 2つ
@@ -90,7 +91,7 @@ export const AllDatePage = ({ posts, checkSignin }: AllDatePage) => {
 
     return (
         <div>
-            <h1>データベース登録</h1>
+            <h1>データベース編集</h1>
             <DropDownVt
                 posts={posts}
                 onVtuberSelect={setSelectedVtuber}
@@ -102,15 +103,15 @@ export const AllDatePage = ({ posts, checkSignin }: AllDatePage) => {
             <DropDownMo
                 posts={posts}
                 selectedVtuber={selectedVtuber}
-                onMovieSelect={setSelectedMovie}
+                onMovieSelect={setSelectedMovie} //このファイルではstringになってる
                 onKaraokeClear={handleMovieClear}
             />
             <DropDownKa
-                posts={posts}
+                posts={posts.karaokes}
                 selectedMovie={selectedMovie}
                 onKaraokeSelect={setSelectedKaraoke}
             />
-            <YoutubePlayer videoId={foundMovie} start={foundKaraokeStart} />
+            <YouTubePlayer videoId={foundMovie} start={foundKaraokeStart} />
             <Link href="/"><button>TOPへ</button></Link>
             <CreateForm
                 posts={posts}
@@ -122,8 +123,236 @@ export const AllDatePage = ({ posts, checkSignin }: AllDatePage) => {
     )
 };
 ///////////////////////////////////////////////////////////////////////////////////////////////////
+
+type selectedDate = {
+    // alljoindata: AllJoinData[];
+    posts: TopPagePosts;
+    selectedVtuber: number;
+    selectedMovie: string;
+    selectedKaraoke: number;
+}
+
+export function CreateForm({ posts, selectedVtuber, selectedMovie, selectedKaraoke }: selectedDate) {
+    const router = useRouter()
+    var defaultValues: CrudDate = {
+        VtuberId: selectedVtuber,//入力不可とする
+        VtuberName: "",
+        VtuberKana: "",
+        IntroMovieUrl: "",
+        MovieUrl: "",
+        MovieTitle: "",
+        KaraokeId: selectedKaraoke,//入力不可とする
+        SingStart: "",
+        SongName: "",
+    }
+
+    // type  CrudDate:
+    const foundVtuber = posts?.vtubers?.find(vtuber => vtuber.VtuberId === selectedVtuber);
+    const foundMovie = posts?.movies?.find(movie => movie.MovieUrl === selectedMovie);
+    const foundMovies = posts?.karaokes?.filter(karaoke => karaoke.MovieUrl === selectedMovie);
+    const foundKaraoke = foundMovies?.find(foundMovie => foundMovie.KaraokeId === selectedKaraoke)
+    console.log("selectedVtuber=", selectedVtuber)
+    console.log("selectedMovie=", selectedMovie)
+    console.log("selectedKaraoke", selectedKaraoke);
+    console.log("foundVtuber", foundVtuber);
+    console.log("foundMovie", foundMovie);
+    console.log("foundMovies", foundMovies);
+    console.log("foundKaraoke", foundKaraoke);
+
+    const [vtuberNameInput, setVtuberNameInput] = useState(foundVtuber?.VtuberName);
+    const [VtuberKanaInput, setVtuberKanaInput] = useState(foundVtuber?.VtuberKana);
+    const [IntroMovieUrInput, setIntroMovieUrInput] = useState(foundVtuber?.IntroMovieUrl);
+    const [MovieUrlInput, setMovieUrlInput] = useState(foundMovie?.MovieUrl);
+    const [MovieTitleInput, setMovieTitleInput] = useState(foundMovie?.MovieTitle);
+    const [SingStartInput, setSingStartInput] = useState(foundKaraoke?.SingStart);
+    const [SongNameInput, setSongNameInput] = useState(foundKaraoke?.SongName);
+
+    const [crudContentType, setCrudContentType] = useState<string>("")
+
+    type EditVtuber = {
+        VtuberId: number;
+        VtuberName: string | null;
+        VtuberKana: string | null;
+        IntroMovieUrl: string | null;
+    }
+    type EditMovie = {
+        VtuberId: number;
+        MovieTitle: string | null;
+        MovieUrl: string;
+    }
+    type EditKaraoke = {
+        MovieUrl: string;
+        KaraokeId: number;
+        SongName: string | null;
+        SingStart: string | null;
+    }
+
+    const axiosClient = axios.create({
+        baseURL: "https://localhost:8080",
+        withCredentials: true,
+        headers: {
+            'Content-Type': 'application/json'
+        },
+    });
+
+    const { register, handleSubmit, formState: { errors }, resetField } = useForm<CrudDate>();
+
+    const onSubmit = async (CrudData: CrudDate) => {
+        console.log("決定押下")
+        console.log("CrudData", CrudData);
+        console.log("choiceCrudType=", crudContentType, "\n selectedVtuber=",
+            selectedVtuber, "\n selectedKaraoke", selectedKaraoke);
+        if (crudContentType === "vtuber" && foundVtuber && selectedVtuber) {
+            try {
+                const reqBody: EditVtuber = {
+                    VtuberId: selectedVtuber,    //変更不可
+                    VtuberName: CrudData.VtuberName || foundVtuber.VtuberName,
+                    VtuberKana: CrudData.VtuberKana || foundVtuber.VtuberKana,
+                    IntroMovieUrl: CrudData.IntroMovieUrl || foundVtuber.IntroMovieUrl,
+                };
+                const response = await axiosClient.post("/edit/vtuber", reqBody);
+                if (!response.status) {
+                    throw new Error(response.statusText);
+                }
+            } catch (err) {
+                console.error(err);
+            }
+        } else if (crudContentType === "movie" && foundMovie && selectedMovie) {
+            try {
+                // const VtuberId =selectedVtuber
+                const reqBody: EditMovie = {
+                    VtuberId: selectedVtuber,                             //変更不可
+                    MovieTitle: CrudData.MovieTitle || foundMovie?.MovieTitle,
+                    MovieUrl: selectedMovie,  //変更不可
+                };
+                const response = await axiosClient.post("/edit/movie", reqBody);
+                if (!response.status) {
+                    throw new Error(response.statusText);
+                }
+            } catch (err) {
+                console.error(err);
+            }
+        } else if (crudContentType === "karaoke" && foundKaraoke && selectedMovie && selectedKaraoke) {
+            try {
+                const reqBody: EditKaraoke = {
+                    MovieUrl: selectedMovie,      //変更不可
+                    KaraokeId: selectedKaraoke,    //変更不可
+                    SongName: CrudData.SongName || foundKaraoke?.SongName,
+                    SingStart: CrudData.SingStart || foundKaraoke?.SingStart,
+                };
+                const response = await axiosClient.post("/edit/karaoke", reqBody);
+                if (!response.status) {
+                    throw new Error(response.statusText);
+                }
+            } catch (err) {
+                console.error(err);
+            }
+            console.log("selectedMovie=", selectedMovie)
+        } else {
+            console.log("編集するデータの種類(vtuber, movie, karaoke)の選択で想定外のエラーが発生しました。")
+        }
+    };;
+
+    return (
+        <div>
+            <button type="button" onClick={() => setCrudContentType("vtuber")}
+            >
+                ＜VTuberを編集＞</button>
+            <button type="button" onClick={() => setCrudContentType("movie")} >
+                ＜歌枠動画を編集＞</button>
+            <button type="button" onClick={() => setCrudContentType("karaoke")} >
+                ＜歌を編集＞</button>
+            <br /><br />
+            {
+                crudContentType === "vtuber" &&
+                <div>
+                    VTuberを編集します。<br />
+                </div>
+            }
+            {
+                crudContentType === "movie" &&
+                <div>
+                    VTuber：{foundVtuber?.VtuberName}<br />
+                    &nbsp;&nbsp; の歌枠動画を編集します。
+                </div>
+            }
+            {
+                crudContentType === "karaoke" &&
+                <div>
+                    VTuber：{foundVtuber?.VtuberName && foundVtuber?.VtuberName}<br />
+                    歌枠動画：{foundMovie?.MovieTitle}
+                    <br />
+                    &nbsp;&nbsp; の歌と開始時間を編集します。
+                </div>
+            }<br />
+            <form onSubmit={handleSubmit(onSubmit)}>
+                {crudContentType === "vtuber" &&
+                    <div>
+                        空欄にした場合、既存データが維持されます<br />
+                        ★Vtuber: <br />
+                        &nbsp;&nbsp;VTuber名:
+                        <input {...register("VtuberName", { required: false })} placeholder={foundVtuber?.VtuberName || "例:妹望おいも"}
+                            // value={foundVtuber?.VtuberName}
+                            onChange={e => setVtuberNameInput(e.target.value)}
+                        /><br />
+                        &nbsp;&nbsp;読み(kana):
+                        <input {...register("VtuberKana", { required: false })} placeholder={foundVtuber?.VtuberKana || "例:imomochi_oimo"}
+                            // value={foundVtuber?.VtuberKana}
+                            onChange={e => setVtuberKanaInput(e.target.value)}
+                        /><br />
+                        {/* {errors.VtuberName && "Vtuber is required"} */}
+                        &nbsp;&nbsp;紹介動画URL(時間指定可):
+                        <input {...register("IntroMovieUrl", { required: false })} placeholder={foundVtuber?.IntroMovieUrl || "例:www.youtube.com/watch?v=AlHRqSsF--8"}
+                            // value={foundVtuber?.IntroMovieUrl || ""}
+                            onChange={e => setIntroMovieUrInput(e.target.value)}
+                        /><br />
+                        {/* {errors.VtuberName && "Vtuber is required"} */}
+                    </div>}
+                {crudContentType === "movie" &&
+                    <div>
+                        ★歌枠動画: <br />
+                        &nbsp;&nbsp;タイトル：
+                        <input {...register("MovieTitle", { required: false })} placeholder={foundMovie?.MovieTitle || "動画タイトル"}
+                            // value={foundMovie?.MovieTitle || ""} 
+                            onChange={e => setMovieTitleInput(e.target.value)}
+                        /><br />
+                        {/* {errors.MovieTitle && "Movie is required"} */}
+                        &nbsp;&nbsp;URL:歌枠動画のURLの変更はできません。
+                        <br />
+                        {/* {errors.MovieUrl && "Url is required"} */}
+                    </div>}
+                {crudContentType === "karaoke" &&
+                    <div>
+                        ★歌: <br />
+                        &nbsp;&nbsp;曲名：
+                        <input {...register("SongName", { required: false })} placeholder={foundKaraoke?.SongName || "曲名"}
+                            // value={foundKaraoke?.SongName || ""}
+                            onChange={e => setSongNameInput(e.target.value)}
+                        /><br />
+                        {/* {errors.SongName && "Song is required"} */}
+                        &nbsp;&nbsp;開始時間:
+                        <input {...register("SingStart", { required: false })} placeholder={foundKaraoke?.SingStart || "例 00:05:30"}
+                            // value={foundKaraoke?.SingStart}
+                            onChange={e => setSingStartInput(e.target.value)}
+                        /><br />
+                        {/* {errors.SingStart && "SingStart is required"} */}
+                    </div>}
+                {/* ※ページ最上部のリストから選択後、✖でクリアすることで、入力フォームを編集できるようになります。<br/> */}
+                <button type="submit"
+                // disabled={status === 'submitting'}
+                >＜決定＞</button>
+            </form>
+            <br /> &nbsp;
+        </div >
+    );
+}
+
+
+
 export async function getServerSideProps(context: { req: { headers: { cookie: any; }; }; }) {
-    const httpsAgent = new https.Agent({ rejectUnauthorized: false });
+    const httpsAgent = new https.Agent({
+        rejectUnauthorized: false
+    });
     let resData;
     try {
         const response = await axios.get(`${domain.backendHost}/vcontents/getalldata`, {
@@ -150,205 +379,3 @@ export async function getServerSideProps(context: { req: { headers: { cookie: an
 }
 
 export default AllDatePage;
-
-/////////////////////////////////////////////////////////////////////////////////////
-type selectedDate = {
-    // alljoindata: AllJoinData[];
-    posts: TopPagePosts;
-    selectedVtuber: number;
-    selectedMovie: string;
-    selectedKaraoke: number;
-}
-
-export function CreateForm({ posts, selectedVtuber, selectedMovie, selectedKaraoke }: selectedDate) {
-    const router = useRouter()
-    const foundVtuber = posts?.vtubers?.find(vtuber => vtuber.Vtuber.VtuberId === selectedVtuber);
-    const foundMovie = posts?.movies?.find(movie => movie.Movie.MovieUrl === selectedMovie);
-    console.log("selectedMovie=", selectedMovie)
-    const foundMovies = posts?.karaokes?.filter(karaoke => karaoke.MovieUrl === selectedMovie);
-    const foundKaraoke = foundMovies?.find(foundMovie => foundMovie.Karaoke.KaraokeId === selectedKaraoke)
-    // console.log("foundKaraoke",foundKaraoke);
-    // console.log("selectedKaraoke",selectedKaraoke);
-
-    const [crudContentType, setCrudContentType] = useState<string>("")
-
-    const [vtuberNameInput, setVtuberNameInput] = useState(foundVtuber?.Vtuber.VtuberName);
-    const [VtuberKanaInput, setVtuberKanaInput] = useState(foundVtuber?.Vtuber.VtuberKana);
-    const [IntroMovieUrInput, setIntroMovieUrInput] = useState(foundVtuber?.Vtuber.IntroMovieUrl);
-    const [MovieUrlInput, setMovieUrlInput] = useState(foundMovie?.Movie.MovieUrl);
-    const [MovieTitleInput, setMovieTitleInput] = useState(foundMovie?.Movie.MovieTitle);
-    const [SingStartInput, setSingStartInput] = useState(foundKaraoke?.SingStart);
-    const [SongNameInput, setSongNameInput] = useState(foundKaraoke?.SongName);
-
-    type CreateVtuber = {
-        VtuberId: number;
-        VtuberName: string;
-        VtuberKana: string;
-        IntroMovieUrl: string | null;
-    }
-    type CreateMovie = {
-        VtuberId: number;
-        MovieTitle: string;
-        MovieUrl: string;
-    }
-    type CreateKaraoke = {
-        MovieUrl: string;
-        KaraokeId: number;
-        SongName: string;
-        SingStart: string;
-    }
-
-    const axiosClient = axios.create({
-        baseURL: "http://localhost:8080",
-        withCredentials: true,
-        headers: {
-            'Content-Type': 'application/json'
-        },
-    });
-
-    const { register, handleSubmit, formState: { errors }, resetField } = useForm<CrudDate>();
-
-    // console.log("vtuberDataForFetch=", vtuberDataForFetch, "\n movieDataForFetch=", movieDataForFetch, "\n karaokeDataForFetch", karaokeDataForFetch)
-    const onSubmit = async (CrudData: CrudDate) => {
-        // console.log("決定押下")
-        // console.log("CrudData", CrudData);
-        if (crudContentType === "vtuber") {
-            try {
-                const reqBody: CreateVtuber = {
-                    VtuberId: 0, //自動振り分け
-                    VtuberName: CrudData.VtuberName,
-                    VtuberKana: CrudData.VtuberKana,
-                    IntroMovieUrl: CrudData.IntroMovieUrl,
-                };
-                const response = await axiosClient.post("/create/vtuber", reqBody);
-                if (!response.status) {
-                    throw new Error(response.statusText);
-                }
-            } catch (err) {
-                console.error(err);
-            }
-        } else if (crudContentType === "movie") {
-            try {
-                const reqBody: CreateMovie = {
-                    VtuberId: selectedVtuber, //既存値
-                    MovieTitle: CrudData.MovieTitle,
-                    MovieUrl: CrudData.MovieUrl,
-                };
-                const response = await axiosClient.post("/create/movie", reqBody);
-                if (!response.status) {
-                    throw new Error(response.statusText);
-                }
-            } catch (err) {
-                console.error(err);
-            }
-        } else if (crudContentType === "karaoke") {
-            try {
-                const reqBody: CreateKaraoke = {
-                    MovieUrl: selectedMovie,//既存値
-                    KaraokeId: 0, //自動振り分け
-                    SongName: CrudData.SongName,
-                    SingStart: CrudData.SingStart,
-                };
-                const response = await axiosClient.post("/create/karaoke", reqBody);
-                if (!response.status) {
-                    throw new Error(response.statusText);
-                }
-            } catch (err) {
-                console.error(err);
-            }
-        } else {
-            console.log("登録するデータの種類(vtuber, movie, karaoke)の選択で想定外のエラーが発生しました。")
-        }
-    };
-
-    return (
-        <div>
-            <button type="button" onClick={() => setCrudContentType("vtuber")} >
-                ＜VTuberを登録＞</button>
-            <button type="button" onClick={() => setCrudContentType("movie")}>
-                ＜歌枠動画を登録＞</button>
-            <button type="button" onClick={() => setCrudContentType("karaoke")}>
-                ＜歌を登録＞</button>
-            <br /><br />
-            {crudContentType === "vtuber" &&
-                <div>
-                    VTuberを登録します。<br />
-                </div>}
-            {crudContentType === "movie" &&
-                <div>
-                    VTuber：{foundVtuber?.Vtuber.VtuberName}<br />
-                    &nbsp;&nbsp; の歌枠動画を登録します。
-                </div>}
-            {crudContentType === "karaoke" &&
-                <div>
-                    VTuber：{foundVtuber?.Vtuber.VtuberName && foundVtuber?.Vtuber.VtuberName}<br />
-                    歌枠動画：{foundMovie?.MovieTitle}
-                    <br />
-                    &nbsp;&nbsp; の歌と開始時間を登録します。
-                </div>}<br />
-            <form onSubmit={handleSubmit(onSubmit)}>
-                {crudContentType === "vtuber" &&
-                    <div>
-                        ★Vtuber: <br />
-                        &nbsp;&nbsp;VTuber名:
-                        <input {...register("VtuberName", { required: true })} placeholder={foundVtuber?.Vtuber.VtuberName || "例:妹望おいも"}
-                            // value={foundVtuber?.VtuberName}
-                            onChange={e => setVtuberNameInput(e.target.value)}
-                        /><br />
-                        &nbsp;&nbsp;読み(kana):
-                        <input {...register("VtuberKana", { required: true })} placeholder={foundVtuber?.Vtuber.VtuberKana || "例:imomochi_oimo"}
-                            // value={foundVtuber?.VtuberKana}
-                            onChange={e => setVtuberKanaInput(e.target.value)}
-                        /><br />
-                        {/* {errors.VtuberName && "Vtuber is required"} */}
-                        &nbsp;&nbsp;紹介動画URL(時間指定可):
-                        <input {...register("IntroMovieUrl", { required: false })} placeholder={foundVtuber?.Vtuber.IntroMovieUrl || "例:www.youtube.com/watch?v=AlHRqSsF--8"}
-                            // value={foundVtuber?.IntroMovieUrl || ""}
-                            onChange={e => setIntroMovieUrInput(e.target.value)}
-                        /><br />
-                        {/* {errors.VtuberName && "Vtuber is required"} */}
-                    </div>}
-                {crudContentType === "movie" &&
-                    <div>
-                        ★歌枠動画: <br />
-                        &nbsp;&nbsp;タイトル：
-                        <input {...register("MovieTitle", { required: true })} placeholder={foundMovie?.MovieTitle || "動画タイトル"}
-                            // value={foundMovie?.MovieTitle || ""} 
-                            onChange={e => setMovieTitleInput(e.target.value)}
-                        /><br />
-                        {/* {errors.MovieTitle && "Movie is required"} */}
-                        &nbsp;&nbsp;URL:
-                        <input {...register("MovieUrl", { required: true })} placeholder={foundMovie?.MovieUrl || "例：www.youtube.com/watch?v=AlHRqSsF--8"}
-                            // value={foundMovie?.MovieUrl}
-                            onChange={e => setMovieUrlInput(e.target.value)}
-                        /><br />
-                        {/* {errors.MovieUrl && "Url is required"} */}
-                    </div>}
-                {crudContentType === "karaoke" &&
-                    <div>
-                        ★歌: <br />
-                        &nbsp;&nbsp;曲名：
-                        <input {...register("SongName", { required: true })} placeholder={foundKaraoke?.SongName || "曲名"}
-                            // value={foundKaraoke?.SongName || ""}
-                            onChange={e => setSongNameInput(e.target.value)}
-                        /><br />
-                        {/* {errors.SongName && "Song is required"} */}
-                        &nbsp;&nbsp;開始時間:
-                        <input {...register("SingStart", { required: true })} placeholder={foundKaraoke?.SingStart || "例 00:05:30"}
-                            // value={foundKaraoke?.SingStart}
-                            onChange={e => setSingStartInput(e.target.value)}
-                        /><br />
-                        {/* {errors.SingStart && "SingStart is required"} */}
-                    </div>}
-                {/* ※ページ最上部のリストから選択後、✖でクリアすることで、入力フォームを編集できるようになります。<br/> */}
-                <button type="submit" >＜決定＞</button>
-                {/* <button type="submit" onClick={() => onSubmit} >＜決定＞</button>      */}
-                {/* <button onClick={resetTextValue} style={{ margin: "10px", background: "gray", color: "#fff" }}>クリア</button>    */}
-            </form>
-            <br />  &nbsp;  &nbsp;  &nbsp;  &nbsp;
-        </div>
-    );
-}
-
-
-// *****memo******
